The playlist command now has an increased limit
Previously it stuck by the YouTube API limit of 50 which meant that even if you imported a playlist with 400 videos, only the first 50 would ne added even if the config had a value bigger than 50. Now, I have added logic in that will allow you to extend that limit. To extend the limit you must increase the paginateMaxLength value in the config. NOTE: For every 50 videos, 1 unit of your API quota will be consumed. That is why I have set the default to 50.