Added CmdRequirementError error class
This error is designed to be thrown when a user does not meet command requirements. Throwing this error whilst processing a command will stop further execution of the command, tell the user on Discord the error message and will not log to the console. However, any other type of error that is thrown will log to the console, and the user will instead be given a pre-written general response informing them an error has occurred.