New option to forget items on finish
This new option makes it so that you can pick between forgetting queue items when they finish, or incrementing a pointer and keeping them. You can change this option in the config. Eventually, I will create a command that encapsulates all options but at the moment they are defined in the config